Explicitly request AMP document when fetching inner app shell component
In the original app shell implementation proposed in ampproject/amp-wp#1519
the is_amp_endpoint() function was modified in order to return true if
an inner component was requested.

In order to achieve the same effect without altering the original is_amp_endpoint()
implementation, we're explicitly asking for an AMP document when requesting
the inner component. Simply, an AMP slug is appended to a URL when fetching
a new page.